Fix Offer and Loan relationship

+++ domain/src/main/kotlin/br/com/amz/replay/vehicle/model/Vehicle.kt
@@ -6,6 +6,6 @@ data class Vehicle(
     val id: UUID? = null,
     val make: String,
     val model: String,
-    val year: Short,
+    val year: Int,
     val milesAmount: Int
 )
